+++ blog/utils/easter_egg.py
